pyproject.toml
[tool.poetry]
name = "sjgeofire"
version = "0.1.0"
description = "GeoFire for Kivy Android apps"
authors = ["Kenechukwu Akubue <kengoon19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyjnius = "^1.6.1"
sjfirebase = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
